opsdroid-bot/requirements.txt
python-jenkins==1.8.0
